[editor] Fixed edit line bug
It used to be that the edit line feature copied the command that you just
entered verbatim into the buffer. It isn't the case, now. It works like insert
line function. This bug was found while manually testing another feature. Just
goes to show you that you have to test everything.