[interpreter] Finished. May need optimizations
But hey, it's just a theory, a ....
Also, some attempts were made at fixing the bug mentioned in the previous commit
where there were memory leaks. At least the double destroying memory error was
found; the underlying leak is still in there somewhere.

This project is turning out to be solely a learning experience, and not much
else.